# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FlavorFind API",
        "description": "FlavorFind is a smart ingredient pairing and recipe discovery API. It answers: 'I have these ingredients — what else goes well with them, and what can I cook?' Powered by 514,780 recipes and a pre-computed ingredient suggestion engine combining co-occurrence analysis and semantic vector similarity.",
        "contact": {
            "name": "FlavorFind Support",
            "email": "dallaherick0@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://flavorfind-22iw.onrender.com",
            "description": "Production"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local development"
        }
    ],
    "paths": {
...
        "/api/v1/recipes/search": {
            "get": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Search recipes by ingredients",
                "description": "Find recipes that contain your ingredients. Two modes: **any** (default) returns recipes with at least one match, ranked by how many ingredients match. **all** returns only recipes that contain every ingredient you listed.",
                "operationId": "606277d69cc0a4ce5191e53ca94b5fae",
                "parameters": [
                    {
                        "name": "ingredient_ids[]",
                        "in": "query",
                        "description": "Ingredient UUIDs to search with (1–15)",
                        "required": true,
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "format": "uuid"
                            }
                        }
                    },
                    {
                        "name": "mode",
                        "in": "query",
                        "description": "'any' = at least one match (default). 'all' = must contain every ingredient.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "any",
                            "enum": [
                                "any",
                                "all"
                            ]
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter by recipe category (partial match)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Chicken"
                        }
                    },
                    {
                        "name": "min_rating",
                        "in": "query",
                        "description": "Minimum recipe rating (1–5)",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "example": 4
                        }
                    },
                    {
                        "name": "max_calories",
                        "in": "query",
                        "description": "Maximum calories per serving",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "example": 500
                        }
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Full-text search on recipe name and description",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "spicy"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Results per page (1–50, default 20)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 20
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number (default 1)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recipe search results",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mode": {
                                            "type": "string",
                                            "example": "any"
                                        },
                                        "results": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/RecipeSummary"
                                            }
                                        },
                                        "pagination": {
                                            "$ref": "#/components/schemas/Pagination"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
        "/api/v1/recipes/{id}": {
            "get": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Get full recipe detail",
                "description": "Returns complete recipe information including step-by-step instructions, full ingredient list with quantities and units, and detailed nutrition data.",
                "operationId": "e20a404de5e2dee076e3ecb5908fb11e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Recipe UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "7db8181c-86ff-4fae-bd4c-998c172237dd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Full recipe detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RecipeDetail"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recipe not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "RecipeSummary": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string",
                        "example": "Chicken Pot Pie"
                    },
                    "category": {
                        "type": "string",
                        "example": "Savory Pies"
                    },
                    "description": {
                        "type": "string"
                    },
                    "prep_time": {
                        "type": "string",
                        "example": "PT30M"
                    },
                    "cook_time": {
                        "type": "string",
                        "example": "PT45M"
                    },
                    "total_time": {
                        "type": "string",
                        "example": "PT1H15M"
                    },
                    "servings": {
                        "type": "integer",
                        "example": 4
                    },
                    "rating": {
                        "type": "number",
                        "format": "float",
                        "example": 4.5
                    },
                    "review_count": {
                        "type": "integer",
                        "example": 128
                    },
                    "calories": {
                        "type": "number",
                        "format": "float",
                        "example": 538.3
                    },
                    "matched_count": {
                        "type": "integer",
                        "example": 2
                    },
                    "total_searched": {
                        "type": "integer",
                        "example": 2
                    }
                },
                "type": "object"
            },
            "RecipeIngredientItem": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string",
                        "example": "chicken breast"
                    },
                    "category": {
                        "type": "string",
                        "example": "protein"
                    },
                    "quantity": {
                        "type": "string",
                        "example": "2"
                    },
                    "quantity_numeric": {
                        "type": "number",
                        "example": 2
                    },
                    "unit": {
                        "type": "string",
                        "example": "lbs"
                    },
                    "sort_order": {
                        "type": "integer",
                        "example": 0
                    }
                },
                "type": "object"
            },
            "RecipeDetail": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string",
                        "example": "Chicken Pot Pie"
                    },
                    "category": {
                        "type": "string",
                        "example": "Savory Pies"
                    },
                    "description": {
                        "type": "string"
                    },
                    "prep_time": {
                        "type": "string",
                        "example": "PT30M"
                    },
                    "cook_time": {
                        "type": "string",
                        "example": "PT45M"
                    },
                    "total_time": {
                        "type": "string",
                        "example": "PT1H15M"
                    },
                    "servings": {
                        "type": "integer",
                        "example": 4
                    },
                    "yield": {
                        "type": "string",
                        "example": "8 servings"
                    },
                    "rating": {
                        "type": "number",
                        "format": "float",
                        "example": 4.5
                    },
                    "review_count": {
                        "type": "integer",
                        "example": 128
                    },
                    "calories": {
                        "type": "number",
                        "format": "float",
                        "example": 538.3
                    },
                    "protein_g": {
                        "type": "number",
                        "format": "float",
                        "example": 32.1
                    },
                    "fat_g": {
                        "type": "number",
                        "format": "float",
                        "example": 18.4
                    },
                    "carbs_g": {
                        "type": "number",
                        "format": "float",
                        "example": 45.2
                    },
                    "fiber_g": {
                        "type": "number",
                        "format": "float",
                        "example": 3.1
                    },
                    "sugar_g": {
                        "type": "number",
                        "format": "float",
                        "example": 5
                    },
                    "sodium_mg": {
                        "type": "number",
                        "format": "float",
                        "example": 820
                    },
                    "instructions": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "keywords": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "ingredients": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RecipeIngredientItem"
                        }
                    }
                },
                "type": "object"
            },
            "Pagination": {
                "properties": {
                    "total": {
                        "type": "integer",
                        "example": 1240
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 20
                    },
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 62
                    }
                },
                "type": "object"
            },
            "Error": {
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Ingredient not found."
                    }
                },
                "type": "object"
            }
        }
    },
...
}